Change client to use pyproject.toml so that dependencies can be managed in a more modern fashion
