+++ client/MANIFEST.in
@@ -1,0 +1,8 @@
+exclude *.egg-info
+
+include README.md
+include LICENSE
+recursive-include src/nv_ingest_client
+include src/version.py
+global-exclude __pycache__
+global-exclude *.pyc
