--- client/requirements.txt
@@ -1,22 +1,0 @@
-azure-storage-blob==12.24.0
-charset-normalizer
-click
-fsspec
-httpx==0.27.2
-llama-index-embeddings-nvidia==0.1.5
-openai==1.40.6
-pydantic>=2.0.0
-pymilvus==2.4.9
-pymilvus[bulk_writer,model]
-pypdfium2
-python-docx
-python-magic
-python-pptx==0.6.23
-redis~=5.0.1
-requests
-setuptools
-tqdm
-langchain-milvus
-langchain-nvidia-ai-endpoints>=0.3.7
-minio
-pyarrow
